refactor: racingGame 파일 분리
